+++ tweet-ui/src/androidTest/java/com/twitter/sdk/android/tweetui/CompactTweetViewTest.java
@@ -19,12 +19,20 @@ package com.twitter.sdk.android.tweetui;
 
 import android.content.Context;
 
+import com.squareup.picasso.Picasso;
+import com.squareup.picasso.RequestCreator;
+
 import com.twitter.sdk.android.core.internal.scribe.EventNamespace;
+import com.twitter.sdk.android.core.models.MediaEntity;
 import com.twitter.sdk.android.core.models.Tweet;
+
+import static org.mockito.Matchers.*;
+import static org.mockito.Mockito.*;
 
 public class CompactTweetViewTest extends BaseTweetViewTest {
     private static final String REQUIRED_TFW_SCRIBE_COMPONENT = "compact";
     private static final String REQUIRED_SDK_SCRIBE_SECTION = "compact";
+    private static final float DELTA = 0.001f;
 
     @Override
     CompactTweetView createView(Context context, Tweet tweet) {
@@ -52,7 +60,7 @@ public class CompactTweetViewTest extends BaseTweetViewTest {
     }
 
     @Override
-    BaseTweetView createViewWithMocks(Context context, Tweet tweet, int styleResId,
+    CompactTweetView createViewWithMocks(Context context, Tweet tweet, int styleResId,
             BaseTweetView.DependencyProvider dependencyProvider) {
         return new CompactTweetView(context, tweet, styleResId, dependencyProvider);
     }
@@ -74,5 +82,48 @@ public class CompactTweetViewTest extends BaseTweetViewTest {
     @Override
     void assertTfwScribeComponent(EventNamespace ns) {
         assertEquals(REQUIRED_TFW_SCRIBE_COMPONENT, ns.component);
+    }
+
+    public void testGetAspectRatio() {
+        assertEquals(1.0, CompactTweetView.getAspectRatio(fakeMediaEntity(100, 100)), DELTA);
+        assertEquals(1.0, CompactTweetView.getAspectRatio(fakeMediaEntity(300, 400)), DELTA);
+        assertEquals(1.0, CompactTweetView.getAspectRatio(fakeMediaEntity(100, 800)), DELTA);
+        assertEquals(1.3333, CompactTweetView.getAspectRatio(fakeMediaEntity(400, 300)), DELTA);
+        assertEquals(1.6666, CompactTweetView.getAspectRatio(fakeMediaEntity(500, 300)), DELTA);
+        assertEquals(2.0, CompactTweetView.getAspectRatio(fakeMediaEntity(600, 300)), DELTA);
+        assertEquals(2.3333, CompactTweetView.getAspectRatio(fakeMediaEntity(700, 300)), DELTA);
+        assertEquals(2.6666, CompactTweetView.getAspectRatio(fakeMediaEntity(800, 300)), DELTA);
+        assertEquals(3.0, CompactTweetView.getAspectRatio(fakeMediaEntity(900, 300)), DELTA);
+        assertEquals(3.0, CompactTweetView.getAspectRatio(fakeMediaEntity(1000, 50)), DELTA);
+    }
+
+    public void testGetAspectRatio_missingSizeDimension() {
+        assertEquals(CompactTweetView.DEFAULT_ASPECT_RATIO,
+                CompactTweetView.getAspectRatio(fakeMediaEntity(0, 0)));
+        assertEquals(CompactTweetView.DEFAULT_ASPECT_RATIO,
+                CompactTweetView.getAspectRatio(fakeMediaEntity(100, 0)));
+        assertEquals(CompactTweetView.DEFAULT_ASPECT_RATIO,
+                CompactTweetView.getAspectRatio(fakeMediaEntity(0, 100)));
+    }
+
+    public void testSetTweetPhoto() {
+        final Picasso mockPicasso = mock(Picasso.class);
+        final RequestCreator mockRequestCreator = mock(RequestCreator.class);
+        MockUtils.mockPicasso(mockPicasso, mockRequestCreator);
+        when(mockDependencyProvider.getImageLoader()).thenReturn(mockPicasso);
+
+        final CompactTweetView tv = createViewWithMocks(context, TestFixtures.TEST_PHOTO_TWEET,
+                R.style.tw__TweetLightStyle, mockDependencyProvider);
+        // assert 2 loads, once for profile photo and once for compact tweet photo
+        verify(mockPicasso, times(2)).load(anyString());
+        // assert fit is called once when the compact tweet photo is loaded
+        verify(mockRequestCreator, times(1)).fit();
+    }
+
+    private MediaEntity fakeMediaEntity(int width, int height) {
+        final MediaEntity.Size medium = new MediaEntity.Size(width, height, "fit");
+        final MediaEntity.Sizes sizes = new MediaEntity.Sizes(null, null, medium, null);
+        return new MediaEntity(null, null, null, 0, 0, 0L, null, null, "fake", sizes, 0L, null,
+                null);
     }
 }