+++ tweet-ui/src/androidTest/java/com/twitter/sdk/android/tweetui/TweetUiTestCase.java
@@ -22,6 +22,7 @@ import android.os.Handler;
 import io.fabric.sdk.android.FabricAndroidTestCase;
 import io.fabric.sdk.android.FabricTestUtils;
 
+import com.squareup.picasso.RequestCreator;
 import com.twitter.sdk.android.core.AppSession;
 import com.twitter.sdk.android.core.Session;
 import com.twitter.sdk.android.core.TwitterApiClient;
@@ -88,8 +89,7 @@ public class TweetUiTestCase extends FabricAndroidTestCase {
     private void createMocks() {
         mainHandler = mock(Handler.class);
         queue = mock(TweetRepositoryTest.TestAuthRequestQueue.class);
-        picasso = mock(Picasso.class);
-        MockUtils.mockPicasso(picasso);
+        picasso = MockUtils.mockPicasso(mock(Picasso.class), mock(RequestCreator.class));
 
         statusesService = mock(StatusesService.class);